requirements.txt
streamlit
openai==0.28.1
langchain==0.0.331
numpy
pandas
PyPDF2
python-dotenv
tiktoken
pypdf
chromadb